svn merge -r 32598:33057 https://dart.googlecode.com/svn/branches/bleeding_edge trunk

git-svn-id: http://dart.googlecode.com/svn/trunk@33060 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/lib/object.cc
@@ -87,7 +87,7 @@ DEFINE_NATIVE_ENTRY(Object_noSuchMethod, 6) {
       }
     }
     if (!function.IsNull()) {
-      const int total_num_parameters = function.NumParameters();
+      const intptr_t total_num_parameters = function.NumParameters();
       const Array& array = Array::Handle(Array::New(total_num_parameters - 1));
       // Skip receiver.
       for (int i = 1; i < total_num_parameters; i++) {
